--- fw18_0279-day4/Search.html
@@ -1,131 +1,0 @@
-
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0/css/all.min.css">
-    <style>
-        #top{
-            height: 120px;
-            width: 75%;
-            margin: auto;
-            margin-bottom: 20px;
-            display: flex;
-        }
-        #im{
-           border: 4px solid rgb(7, 167, 7); 
-           width: 60%;
-           display: flex;
-        }
-        #im img{
-            height: 100px;
-            width: 100px;
-            padding: 5px;
-        }
-        #im>div{
-            margin-left: 20px;
-            font-family: calibri;
-        }
-        #tag h2{
-            margin-bottom: -10px;
-        }
-        #ri{
-              background-color: #EDF1E9;
-              width: 300px;
-              padding: 20px;
-              margin-left: 20px;
-             
-        }
-        i{
-            font-weight: bold;
-            font-family: calibri;
-            font-size: 21px;
-            display: block;
-           margin-bottom: 20px;
-        }
-        #btn{
-            background-color: blue;
-            color: white;
-            border: none;
-            height: 34px;
-            padding: 3px;
-            width: 55px;
-            cursor: pointer;
-           border-radius: 5px;
-           padding-top: 5px;
-        }
-        .btn{
-            margin-left: 250px;
-            margin-top: -35px;
-            
-            width: 55px;
-        }
-        #ri>input{
-            height: 28px;
-            width: 220px;
-        }
-    /*  ----------------         For Searching          ------------------      */
-   
-
-    hr{
-        margin: auto;
-        width: 75%;
-        margin-top: -10px;
-        margin-bottom: 20px;
-    }
-    #Search{
-        height: auto;
-        width: 80%;
-      
-        margin: auto;
-       }
-     
-       img{
-        height: 120px;
-        width: 150px;
-        
-        
-       }
-       h1{
-        color: blue;
-        font-family: arial;
-        font-size: 15px;
-        display: block;
-        width:500px;
-        
-        
-       }
-       h5{
-        
-        color: blue;
-        font-family: arial;
-        
-       }
-     #K{
-        height:120px ;
-        cursor: pointer;
-        border: 1px solid black;
-        margin-left: 10px;
-     }
-     #div{
-        display: flex;
-        margin-bottom: 20px;
-     }
-    </style>
-</head>
-<body>
-    <div id="top">
-      
-    </div>
-    <!--*********** #region  second part **************-->
-
-    <h3 style="margin-left:12% ;font-family: arial;">Browse Common Foods</h3>
-    <hr>
-    <div id="Search" ></div>
-   
-</body>
-</html>
-<script src="./scripts/Search.js" type="module"></script>
