--- fw18_0279-day4/food.html
@@ -1,132 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
-     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0/css/all.min.css">
-     
-    <style>
-       #top{
-            height: 120px;
-            width: 75%;
-            margin: auto;
-            margin-bottom: 20px;
-            display: flex;
-        }
-        #im{
-           border: 4px solid rgb(7, 167, 7); 
-           width: 60%;
-           display: flex;
-        }
-        #im img{
-            height: 100px;
-            width: 100px;
-            padding: 5px;
-        }
-        #im>div{
-            margin-left: 20px;
-            font-family: calibri;
-        }
-        
-      /* ********* ----  Another Side Part  ---------- ******** */
-        #ri{
-              background-color: #EDF1E9;
-              width: 300px;
-              padding: 20px;
-              margin-left: 20px;
-             
-        }
-       
-        #tag p{
-            margin-left: -240px;
-            margin-top: 60px;
-        }
-        i{
-            font-weight: bold;
-            font-family: calibri;
-            font-size: 21px;
-            display: block;
-           margin-bottom: 20px;
-        }
-        #btn{
-            background-color: blue;
-            color: white;
-            border: none;
-            height: 34px;
-            padding: 3px;
-            width: 55px;
-            cursor: pointer;
-           border-radius: 5px;
-           padding-top: 5px;
-           margin-top: -85px;
-        }
-        .btn{
-            margin-left: 250px;
-            margin-top: -35px;
-            
-            width: 55px;
-        }
-        #ri>input{
-            height: 28px;
-            width: 220px;
-            margin-left: -300px;
-            margin-top: 30px;
-        }
-      /* -------------  bottom   ----------------*/
-
-       #hom{
-        height: auto;
-        width: 80%;
-      
-        margin: auto;
-       }
-       div{
-        display: grid;
-        grid-template-columns: repeat(2,1fr);
-        grid-template-rows: repeat(5,150px);
-        
-       }
-       img{
-        height: 120px;
-        width: 150px;
-        
-       }
-       h1{
-        color: blue;
-        font-family: arial;
-        font-size: 20px;
-        display: block;
-        width:300px;
-        
-        
-       }
-       h5{
-        
-        color: blue;
-        font-family: arial;
-        
-       }
-      #K{
-        height: auto;
-        width: 60%;
-        margin-left: -20%;
-        margin-top: -5px;
-        display: block;
-        cursor: pointer;
-      }
-       
-        
-    </style>
-</head>
-<body>
-     <div id="top">
-        
-      
-    </div>
-    <div id="hom"></div>
-    
-</body>
-</html>
-<script src="./scripts/food.js" type="module"></script>
