+++ fw18_0279-day5/components/infonav.js
@@ -1,0 +1,18 @@
+let nav2 = () => {
+    return `<div id="im">
+            <img src="https://img.freepik.com/free-vector/spoon-fork-knife-plate-3d-illustration_33099-205.jpg" alt="">
+            <div id="tag">
+                <h2>Foods</h2>
+                <p>Browse common foods and products from your favourite brands and restaurants.<p>
+            </div>
+        </div>
+        <!-- #region right search bar-->
+        <div id="ri">
+            <i>Food Search</i>
+            <input type="text" placeholder="Enter search" id="query">
+            <div class="btn">
+            <button id="btn" ><i class="fa-solid fa-magnifying-glass"></i></button>
+            </div>
+        </div>`
+}
+export {nav2}
